backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
motor==3.6.0
openai>=1.0.0
python-dotenv==1.0.0
python-multipart==0.0.9
PyPDF2==3.0.1
